Calculator based on basic C (KOSS)
This program depicts a simple calculator based on basic C language.
Operations such as division, multiplication, addition, subtraction and power are possible following the rule of BODMAS.
But, as brackets are not processed, it follows DMAS strictly. The user can input decimal values as well.
Example:-
Input:-
56+2*35/7
Output:-
5.000000
10.000000
66.000000

The last output is the Final Answer.